+++ cinemadb/src/main/java/com/cinemabookingsystem/cinemadb/service/TicketPriceService.java
@@ -1,12 +1,10 @@
 package com.cinemabookingsystem.cinemadb.service;
 
-import java.math.BigDecimal;
-
 import com.cinemabookingsystem.cinemadb.dto.TicketPriceDTO;
 import com.cinemabookingsystem.cinemadb.model.TicketPrice;
 import com.cinemabookingsystem.cinemadb.model.TicketType;
 
 public interface TicketPriceService {
-    public BigDecimal getPriceByTicketType(TicketType ticketType);
+    public Float getPriceByTicketType(TicketType ticketType);
     public TicketPrice setTicketPrice(TicketPriceDTO TicketPriceDTO);
 }
